+++ .config/wget/wgetrc.default.sh
@@ -2,6 +2,10 @@
 
 DIRNAME="$(dirname "$0")"
 
-tee "$DIRNAME"/wgetrc >/dev/null <<EOF
+if [ -f "$DIRNAME"/wgetrc ]; then
+    echo "$DIRNAME"/wgetrc already exists
+else
+    tee "$DIRNAME"/wgetrc >/dev/null <<EOF
 hsts-file = ${XDG_DATA_HOME}/wget/wget-hsts
 EOF
+fi
